[MH3-24596] - Added UDP UDP_OPTION_ENFORCE_STRICT_NET option. Set the network interface for the DHCP server transfers. Updated the interface mapping. Updated statistics.

+++ tcpip/src/tcpip_manager_control.h
@@ -288,9 +288,8 @@ typedef struct _TCPIP_STACK_MODULE_CTRL
     // i.e., across StackInit() -> StackDeInit() calls
     // 
     //
-    // number of the interfaces supported in this session
-    int     nIfs;         
-    int     nAliases;       // number of alias interfaces in this session         
+    uint16_t    nIfs;       // number of the interfaces supported in this session
+    uint16_t    nAliases;   // number of alias interfaces in this session         
 	// number of the modules enabled in this session
 	int 	nModules;
     // allocation parameters
